CP-306_JNI_&_INIT

+++ p7libapi/src/main/java/ru/petroplus/pos/p7LibApi/IP7LibCallbacks.kt
@@ -67,7 +67,7 @@ interface IP7LibCallbacks {
      * @param cardNumber - номер карты
      * @param transactionEntity - <=> Запись в базе данных
      */
-    fun findLastTransactionDB(cardNumber: UInt, record: TransactionRecordDto): ResultCode
+    fun findLastTransactionDB(cardNumber: Long, record: TransactionRecordDto): ResultCode
 
     /**
      * Обновление записей транзакции в БД
