+++ p7libapi/src/main/java/ru/petroplus/pos/p7LibApi/dto/TransactionInfoDto.kt
@@ -1,7 +1,6 @@
 package ru.petroplus.pos.p7LibApi.dto
 
 /**
- * TODO -  Юрий добавь описание
  * @property transNumber - номер последней транзауции
  * @property cardNumber - номер карты
  * @property terminalNumber - номер терминала
@@ -9,9 +8,9 @@ package ru.petroplus.pos.p7LibApi.dto
  * @property issuerId - ID эмитента
  */
 class TransactionInfoDto(
-    var transNumber:    UInt = 0u,
-    var cardNumber:     UInt = 0u,
-    var terminalNumber: UInt = 0u,
-    var terminalId:     UInt = 0u,
-    var issuerId:       UInt = 0u
+    var transNumber:    Long = 0,
+    var cardNumber:     Long = 0,
+    var terminalNumber: Long = 0,
+    var terminalId:     Long = 0,
+    var issuerId:       Long = 0
 )
